requirements.txt
Django==5.2.11
djangorestframework==3.15.2
django-filter==24.3
celery==5.3.4
redis==5.0.1
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
watchdog==3.0.0
